+++ libs/analysis/bbhnet/analysis/ledger/events.py
@@ -1,0 +1,84 @@
+from dataclasses import dataclass
+from typing import List, TypeVar
+
+import numpy as np
+
+from bbhnet.analysis.ledger.injections import InterferometerResponseSet
+from bbhnet.analysis.ledger.ledger import Ledger, metadata, parameter
+
+SECONDS_IN_YEAR = 31556952
+F = TypeVar("F", np.ndarray, float)
+
+
+@dataclass
+class TimeSlideEventSet(Ledger):
+    Tb: float = metadata(0)
+    decision_statistic: np.ndarray = parameter()
+    time: np.ndarray = parameter()
+
+    def compare_metadata(self, key, ours, theirs):
+        if key == "Tb":
+            return ours + theirs
+        return super().compare_metadata(key, ours, theirs)
+
+    def nb(self, threshold: F) -> F:
+        try:
+            repeats = len(threshold)
+        except TypeError:
+            return (self.decision_statistic >= threshold).sum()
+        else:
+            stats = self.decision_statistic[:, None]
+            stats = np.repeat(stats, repeats, axis=1)
+            return (stats >= threshold).sum(0)
+
+    def far(self, threshold: F) -> F:
+        nb = self.nb(threshold)
+        return SECONDS_IN_YEAR * nb / self.Tb
+
+    def significance(self, threshold: F, T: float) -> F:
+        """see https://arxiv.org/pdf/1508.02357.pdf, eq. 17
+        Represents the likelihood that at least one event
+        with detection statistic value `threshold` will occur
+        after observing this distribution for a period `T`.
+        Args:
+            threshold: The detection statistic to compare against
+            T:
+                The length of the analysis period in which the
+                detection statistic was measured, in seconds
+        """
+
+        nb = self.nb(threshold)
+        return 1 - np.exp(-T * (1 + nb) / self.Tb)
+
+
+@dataclass
+class EventSet(TimeSlideEventSet):
+    shift: np.ndarray = parameter()
+
+    @classmethod
+    def from_timeslide(cls, event_set: TimeSlideEventSet, shift: List[float]):
+        shifts = np.ndarray([shift] * len(event_set))
+        d = {k: getattr(event_set, k) for k in event_set.__dataclass_fields__}
+        return cls(shift=shifts, **d)
+
+
+# inherit from TimeSlideEventSet since injection
+# will already have shift information recorded
+@dataclass
+class RecoveredInjectionSet(TimeSlideEventSet, InterferometerResponseSet):
+    @classmethod
+    def recover(
+        cls,
+        events: TimeSlideEventSet,
+        responses: InterferometerResponseSet,
+        offset: float,
+    ):
+        # TODO: need an implementation that will
+        # also do masking on shifts
+        diffs = np.abs(events.time - responses.gps_time[:, None] - offset)
+        idx = diffs.argmin(axis=-1)
+        events = events[idx]
+        kwargs = events.__dict__ | responses.__dict__
+        for attr in responses.waveform_fields:
+            kwargs.pop(attr)
+        return cls(**kwargs)
